fix wrong encode for body with basic types
